update file tree
